# Precompile some functions for faster startup