See discussion in https://github.com/JuliaLang/julia/issues/12897