@timholy 2015-09-02
(Test failure on master is due to https://github.com/simonster/Reexport.jl/pull/3).