@dcjones 2015-09-07
I compared this to @spencerlyon2's patch in #672, and this one does better so I'm merging it, but surprisingly it's only a little better (`runtests.jl` takes 130 seconds, versus 145 seconds in #672).

It's fun to see Gadfly creep towards not being the slowest plotting software ever written. :grinning: