@timholy 2015-09-08
It might get better still if I get around to submitting my precompiles to Iterators, DataStructures, and DataFrames

> It's fun to see Gadfly creep towards not being the slowest plotting software ever written. :grinning:

While performance continues to be a challenge in some areas, it's clearly much better than when I last looked. And some of the issues are due to limitations in julia itself. As I dig further into the code, I have to say I grow more and more fond of the architecture. You've been able to keep a fairly clean design over ~20k loc (including Compose) for something that has to handle a huge number of cases and messy user input. No small feat. I just hope that the performance improvements I submit don't uglify things too much.