@timholy 2015-09-08
A package I'll soon try to polish, document, tag, and release: https://github.com/timholy/SnoopCompile.jl