@spencerlyon2 2015-09-08
That's [sneaky](https://github.com/timholy/SnoopCompile.jl/blob/master/src/snoop.patch) and great!